feat: components and css refactor
