+++ nerdlets/geo-ops-nerdlet/DefineLocations/DefineLocations.js
@@ -1,7 +1,6 @@
 /* eslint-disable react/no-did-update-set-state */
 import React from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import cloneDeep from 'lodash.clonedeep';
 import { Button, navigation, Spinner, Stack, StackItem } from 'nr1';
 
@@ -9,64 +8,17 @@ import {
   DEFINE_LOCATIONS_UI_SCHEMA,
   MAP_LOCATION_JSON_SCHEMA,
   MAP_LOCATION_DEFAULTS
-} from '../shared/constants';
+} from '../../shared/constants';
 
-import { JsonSchemaForm, LocationTable } from '../shared/components';
-import { writeMapLocation } from '../shared/services/map-location';
+import { LocationTable } from '../../shared/components';
+import { writeMapLocation } from '../../shared/services/map-location';
 
-const FileUploadContainer = styled.div`
-  text-align: center;
-
-  p {
-    text-align: center;
-    color: #8e9494;
-  }
-`;
-
-const ManualDescription = styled.div`
-  h4 {
-    text-align: center;
-    margin-bottom: 4px;
-  }
-
-  p {
-    text-align: center;
-    color: #8e9494;
-  }
-`;
-
-const OrLine = styled.div`
-  display: flex;
-  height: 1px;
-  justify-content: center;
-  align-items: center;
-  margin: 36px 0;
-  border: none;
-  overflow: visible;
-  border-top: 1px solid #e3e4e4;
-
-  &:before {
-    content: 'or';
-    background-color: #fff;
-    padding: 0 8px;
-    text-align: Center;
-    position: relative;
-    bottom: 2px;
-    color: #8e9494;
-    font-style: italic;
-  }
-`;
-
-const StyledJsonSchemaForm = styled(JsonSchemaForm)`
-  padding: 24px;
-  border-radius: 4px;
-  background-color: #fafbfb;
-  border: 1px solid #e3e4e4;
-
-  .form-group {
-    margin-bottom: 16px;
-  }
-`;
+import {
+  FileUploadContainer,
+  ManualDescription,
+  OrLine,
+  StyledJsonSchemaForm
+} from './styles';
 
 export default class DefineLocations extends React.PureComponent {
   static propTypes = {
