+++ nerdlets/geo-ops-nerdlet/EmptyState/EmptyState.js
@@ -1,69 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Grid, GridItem, Stack, StackItem, Button } from 'nr1';
+import { Grid, Stack, Button } from 'nr1';
 import { Map, TileLayer } from 'react-leaflet';
 import { EmptyState } from '@newrelic/nr1-community';
-import styled from 'styled-components';
-import { ToolbarWrapper, ToolbarItem } from '../shared/components';
+import { ToolbarWrapper } from '../../shared/components';
 
-const StyledToolbarItem = styled(ToolbarItem)`
-  font-size: 1.25em;
-  color: #8e9494;
-`;
-
-const StyledGrid = styled(Grid)`
-  height: calc(100% - 74px);
-`;
-
-const LeftGridItem = styled(GridItem)`
-  overflow: scroll;
-  overflow-x: hidden;
-  background-color: #fff;
-  box-shadow: 0px 1px 0px rgba(0, 75, 82, 0.17),
-    0px 3px 2px rgba(0, 49, 54, 0.05), 0px 1px 3px rgba(0, 134, 148, 0.08),
-    0px 8px 17px rgba(88, 114, 117, 0.25);
-`;
-
-const RightGridItem = styled(GridItem)`
-  position: relative;
-`;
-
-const MapWrapper = styled.div`
-  width: 100%;
-  height: 100%;
-`;
-
-const GetStartedWrapper = styled(Stack)`
-  display: flex;
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: 1000;
-  backdrop-filter: blur(10px);
-`;
-
-const GetStartedPopover = styled(StackItem)`
-  max-width: 440px;
-  text-align: center;
-  padding: 24px;
-  background-color: #fff;
-  border-radius: 4px;
-  box-shadow: 0px 1px 0px rgba(0, 75, 82, 0.11),
-    0px 3px 0px rgba(0, 49, 54, 0.04), 0px 1px 3px rgba(0, 134, 148, 0.03),
-    0px 8px 7px rgba(70, 107, 111, 0.05);
-
-  p {
-    color: #8e9494;
-  }
-
-  hr {
-    margin: 20px 0;
-    border: none;
-    border-top: 1px dotted #e3e4e4;
-  }
-`;
+import {
+  StyledToolbarItem,
+  StyledGrid,
+  LeftGridItem,
+  RightGridItem,
+  MapWrapper,
+  GetStartedWrapper,
+  GetStartedPopover
+} from './styles';
 
 const EmptyStateWrapper = ({ navigation }) => {
   const STARTING_CENTER = [39.5, -98.35];
