+++ nerdlets/geo-ops-nerdlet/index.js
@@ -4,7 +4,7 @@ import { navigation, NerdletStateContext, Spinner } from 'nr1';
 
 import cloneDeep from 'lodash.clonedeep';
 
-import EmptyState from './EmptyState';
+import EmptyState from './EmptyState/EmptyState';
 import CreateMap from './CreateMap';
 import ViewMap from './ViewMap/ViewMap';
 import MapList from './MapList/MapList';
